Add separate manager project
