+++ manager/src/main/scala/org/labrad/manager/Handlers.scala
