+++ manager/src/main/scala/org/labrad/manager/Hub.scala
