+++ manager/src/main/scala/org/labrad/manager/Listener.scala
