+++ manager/src/main/scala/org/labrad/manager/LoginHandler.scala
