+++ manager/src/main/scala/org/labrad/manager/ManagerUtils.scala
