+++ manager/src/main/scala/org/labrad/manager/Messager.scala
