+++ manager/src/main/scala/org/labrad/manager/RemoteConnector.scala
