+++ manager/src/main/scala/org/labrad/manager/auth/AuthServer.scala
