+++ manager/src/main/scala/org/labrad/manager/auth/AuthStore.scala
