+++ manager/src/main/scala/org/labrad/manager/auth/OAuth.scala
