+++ manager/src/main/scala/org/labrad/manager/auth/package.scala
