+++ manager/src/main/scala/org/labrad/registry/DelphiStore.scala
