+++ manager/src/main/scala/org/labrad/registry/FileStore.scala
