+++ manager/src/main/scala/org/labrad/registry/Migrate.scala
