+++ manager/src/main/scala/org/labrad/registry/Registry.scala
