+++ manager/src/main/scala/org/labrad/registry/RemoteStore.scala
