+++ manager/src/main/scala/org/labrad/registry/SQLiteStore.scala
