+++ manager/src/test/scala/org/labrad/ClientTest.scala
