+++ manager/src/test/scala/org/labrad/ManagerTest.scala
