+++ manager/src/test/scala/org/labrad/ReflectTest.scala
