+++ manager/src/test/scala/org/labrad/RegistryTest.scala
