+++ manager/src/test/scala/org/labrad/ServerTest.scala
