+++ manager/src/test/scala/org/labrad/manager/ManagerConfigTest.scala
