+++ manager/src/test/scala/org/labrad/manager/auth/AuthStoreTest.scala
