+++ manager/src/test/scala/org/labrad/registry/RemoteStoreTest.scala
